PerfectNumber.c Add Math Algorithm
This add Perfect Number algorithm which finds the list of all perfect numbers between a given interval. A Perfect number is a positive integer that is equal to the sum of its positive divisors excluding the number itself eg. 6 is a perfect number. As the sum of the divisors (1+2+3 = 6) are equal to the number itself.

Closes https://github.com/fnplus/interview-techdev-guide/issues/133