1.Search.py Add Search operation for BST
Binary Search Tree searches a key before Inserting a new key at the leaf node. The algorithm performs the following. To search a given key, first compares it with root and if the key is present at the root, then we return the root. If key is greater than the root's key, we recur for right subtree of root node. Otherwise we recur for left subtree

Closes https://github.com/fnplus/interview-techdev-guide/issues/165