Add Insertion sort Algorithm - Java
https://en.wikipedia.org/wiki/Insertion_sort

Insertion sort is an algorithm that sorts a list one item at a time. The list is split into a sorted part and non-sorted part. With each iteration, it takes the next value waiting to be sorted (in the unsorted part) and it inserts it into the sorted part, into its proper location within the sorted items.